requirements.txt
aioamqp-beta==0.0.1
aiomysql==0.0.20
aioredis==1.2.0
asn1crypto==0.24.0
async-timeout==3.0.1
cffi==1.12.3
cryptography==2.7
hiredis==1.0.0
motor==2.0.0
pamqp==2.3.0
peewee==2.10.2
peewee-async==0.5.12
pycparser==2.19
pymongo==3.8.0
PyMySQL==0.9.2
six==1.12.0
tornado==6.0.3
websocket-client==0.56.0
